Decompose page component into smaller pieces
